refactor: move auth store to api directory and update import paths
